requirements.txt
# Core data processing and ML libraries
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
joblib>=1.2.0

# Kafka integration
kafka-python>=2.0.0

# Spark for distributed processing
pyspark>=3.3.0

# Network traffic analysis
scapy>=2.4.0

# Logging and utilities
python-dateutil>=2.8.0

# Optional: Jupyter for model development
jupyter>=1.0.0
ipykernel>=6.0.0

# Optional: Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Optional: Additional ML libraries
xgboost>=1.6.0
lightgbm>=3.3.0

# Development and testing
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0
